Use Dart Sass and the embedded compiler from HEAD by default
This matches the behavior of other CI tooling, and is low-risk since
the release process releases the versions from HEAD anyway. It also
eliminates the need to make empty version-only updates to the compiler
just to make logic in Dart Sass visible to the test infrastructure.

This also refactors the tool/ directory to move each individual
`getX()` function into its own file, rather than lumping them all
together in the ambiguously-named `utils.ts`. Any utils functions that
are only used in one file have been moved to that file.
